fix: 修正showtoast error icon

+++ packages/framework-plugin-low-code/src/builder/mp/index.ts
@@ -463,22 +463,24 @@ async function cleanPkg(pkg: IWeAppData, projDir: string) {
  * Delete unsed materials
  */
 function cleanMaterils(materialsDir: string, usedComps: IUsedComps) {
-  fs.readdirSync(materialsDir)
-    .filter(junk.not)
-    .map((libName) => {
-      const libDir = path.join(materialsDir, libName)
+  if (fs.existsSync(materialsDir)) {
+    fs.readdirSync(materialsDir)
+      .filter(junk.not)
+      .map((libName) => {
+        const libDir = path.join(materialsDir, libName)
 
-      if (
-        fs.existsSync(path.join(libDir, 'meta.json')) ||
-        fs.existsSync(path.join(libDir, 'mergeMeta.json'))
-      ) {
-        // Skip none-composited materials
-        return
-      }
-      if (!usedComps[libName]) {
-        removeFile(libDir)
-        return
-      }
-      cleanDir(libDir, [...Array.from(usedComps[libName]), 'libCommonRes'])
-    })
+        if (
+          fs.existsSync(path.join(libDir, 'meta.json')) ||
+          fs.existsSync(path.join(libDir, 'mergeMeta.json'))
+        ) {
+          // Skip none-composited materials
+          return
+        }
+        if (!usedComps[libName]) {
+          removeFile(libDir)
+          return
+        }
+        cleanDir(libDir, [...Array.from(usedComps[libName]), 'libCommonRes'])
+      })
+  }
 }
